requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pypdf==4.2.0
python-multipart==0.0.9
pydantic==2.7.1
google-generativeai==0.7.2
sentence-transformers==3.0.1
numpy==1.26.4
